chapter-10/ml-example/dev_requirements.txt
-r requirements.txt
pytest
pytest-cov
apache-airflow
pre-commit
jupyterlab